requirements.txt
amazon.ion==0.9.1
asgiref==3.5.0
boto3==1.21.40
botocore==1.24.40
Django==4.0.4
djangorestframework==3.13.1
ionhash==1.2.1
jmespath==1.0.0
mysqlclient==2.1.0
pyqldb==3.2.2
python-dateutil==2.8.2
pytz==2022.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
